LayoutTests/media/media-source/media-source-real-webm-sourcebufferlist-events.html

<!DOCTYPE html>
<html>
<head>
    <title>media-source-real-webm-sourcebufferlist-events</title>
    <script src="webm-generator.js"></script>
    <script src="../video-test.js"></script>
    <script>
    // Tests that SourceBufferList fires addsourcebuffer and removesourcebuffer
    // events when SourceBuffers are added/removed.
    // MSE spec section 3.4.3 — event firing and ordering.

    var source;
    var sourceBuffer;

    async function runTest() {
        findMediaElement();

        if (!MediaSource.isTypeSupported(WebM.VIDEO_TYPE)) {
            consoleWrite('SKIPPED: ' + WebM.VIDEO_TYPE + ' is not supported');
            return;
        }

        source = new MediaSource();
        video.src = URL.createObjectURL(source);
        await waitFor(source, 'sourceopen');

        // Listen for addsourcebuffer on sourceBuffers
        var addPromise = waitFor(source.sourceBuffers, 'addsourcebuffer', true);

        sourceBuffer = source.addSourceBuffer(WebM.VIDEO_TYPE);
        testExpected('source.sourceBuffers.length', 1);

        // The event may be synchronous or task-queued; give it a chance to fire
        var result = await Promise.race([
            addPromise.then(() => 'fired'),
            new Promise(resolve => setTimeout(() => resolve('timeout'), 100))
        ]);
        consoleWrite('addsourcebuffer event on sourceBuffers: ' + result);

        // Append init to make track active
        var initData = WebM.initSegment({
            tracks: [{ id: 1, type: 'video' }]
        });
        sourceBuffer.appendBuffer(initData);
        await waitFor(sourceBuffer, 'updateend');

        testExpected('source.activeSourceBuffers.length', 1);

        // Listen for removesourcebuffer
        var removePromise = waitFor(source.sourceBuffers, 'removesourcebuffer', true);

        run('source.removeSourceBuffer(sourceBuffer)');
        testExpected('source.sourceBuffers.length', 0);

        result = await Promise.race([
            removePromise.then(() => 'fired'),
            new Promise(resolve => setTimeout(() => resolve('timeout'), 100))
        ]);
        consoleWrite('removesourcebuffer event on sourceBuffers: ' + result);
    }

    window.addEventListener('load', event => {
        runTest().then(endTest).catch(failTest);
    });
    </script>
</head>
<body>
    <video></video>
</body>
</html>
